requirements.txt
langchain
langgraph
google-genai
streamlit
pyTigerGraph==1.8.0
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
scikit-learn
requests
pillow
pytesseract
httpx
pymupdf
python-docx
tavily-python
groq